Added "Remember Me" unique token system.
This should be way more secure than storing the user's password in a cookie. Using this system, the token lasts a week and changes every time the user logs in or out. This way, the chances of someone stealing the cookie are slim and the user's password is never revealed.